refactor: qc-application creator 도메인 패키지 정리
