+++ qc-application/src/main/java/com/eager/questioncloud/application/api/creator/service/CreatorService.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.creator;
+package com.eager.questioncloud.application.api.creator.service;
 
 import com.eager.questioncloud.core.domain.creator.dto.CreatorInformation;
 import com.eager.questioncloud.core.domain.creator.infrastructure.CreatorRepository;
